requirements.txt
streamlit
langchain
langchain-openai
pandas
python-dotenv
langchain_experimental